Add files via upload
Updated ISR database for assignment of serotype to Salmonella enterica subspecies I. For each ISR, a reference to an NCBI genome is indicated or "NoNCBImatch" is indicated. The term "false" in the heading means that the ISR sequence happens to be identical to a 5S ribosomal region that is not linked to the dkgB gene. In all the cases examined thus far, false matches are at the 5S ribosomal sequence near the hdfR gene.